requirements.txt
jinja2
python-dotenv
langchain
langchain-google-genai
